backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
psycopg==3.1.18
django-environ==0.11.2